# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "PetStore APIs exposed",
    "version" : "1.0.0",
    "title" : "Petstore APIs",
    "contact" : {
      "name" : "Antonio Goncalves",
      "url" : "https://antoniogoncalves.org/",
      "email" : "antonio.goncalves@gmail.com"
    }
  },
  "host" : "localhost:5000",
  "basePath" : "/rest",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/items" : {
      "get" : {
        "tags" : [ "Item" ],
        "summary" : "Lists all items",
        "description" : "",
        "operationId" : "listAllItems",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Item"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Item" ],
        "summary" : "Creates a new item",
        "description" : "",
        "operationId" : "createItem",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Item"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/items/{id}" : {
      "get" : {
        "tags" : [ "Item" ],
        "summary" : "Finds an item by its id",
        "description" : "",
        "operationId" : "findItemById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "Item" ],
        "summary" : "Updates an item",
        "description" : "",
        "operationId" : "updateItem",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Item"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Item" ],
        "summary" : "Deletes an item by its id",
        "description" : "",
        "operationId" : "deleteItemById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/products" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Lists all products",
        "description" : "",
        "operationId" : "listAllProducts",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Product"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Product" ],
        "summary" : "Creates new product",
        "description" : "",
        "operationId" : "createProduct",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Product"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/products/{id}" : {
      "get" : {
        "tags" : [ "Product" ],
        "summary" : "Finds a product by id",
        "description" : "",
        "operationId" : "findProductById",
        "produces" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "Product" ],
        "summary" : "Updates a product",
        "description" : "",
        "operationId" : "updateProduct",
        "consumes" : [ "application/xml", "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Product"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "Product" ],
        "summary" : "Deletes a product by id",
        "description" : "",
        "operationId" : "deleteProductById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int64"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Category" : {
      "type" : "object",
      "required" : [ "description", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 30
        },
        "description" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 3000
        }
      }
    },
...
    "Item" : {
      "type" : "object",
      "required" : [ "description", "name", "unitCost" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 30
        },
        "description" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 3000
        },
        "imagePath" : {
          "type" : "string"
        },
        "unitCost" : {
          "type" : "number",
          "format" : "float"
        },
        "product" : {
          "$ref" : "#/definitions/Product"
        }
      }
    },
    "Product" : {
      "type" : "object",
      "required" : [ "description", "name" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 30
        },
        "description" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 3000
        },
        "category" : {
          "$ref" : "#/definitions/Category"
        }
      }
    }
  }
}